api/pyproject.toml
[tool.poetry]
name = "repast"
version = "0.1.0"
description = ""
authors = ["drewpayment <drew.payment@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
flask = "^3.1.0"
flask-cors = "^5.0.0"


[tool.poetry.group.dev.dependencies]
google-generativeai = "^0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
